feat(auth): implementado sistema de autorização baseado em cargos (aluno, professor e admin) (#10)
* refactor(auth): realocação e renomeação de arquivos e funções / authUsuario removido

* feat(auth): rotas fechadas com sistema de autorização
