+++ src/features/Repertorios/Schemas/ObraSchema.ts
@@ -9,7 +9,11 @@ import {
   idValidation,
 } from '../../../shared/Validations';
 import { genericError, schemaValidationError } from '../../../shared/Schemas';
-import { authMiddleware, optionalAuthMiddleware } from '../../Auth/Plugins';
+import {
+  authMiddleware,
+  optionalAuthMiddleware,
+} from '../../../shared/middlewares/Authentication';
+import { authProfessor } from '../../../shared/middlewares/Authorization';
 
 export const ObraSchema: EntitySchema = {
   get: {
@@ -28,7 +32,7 @@ export const ObraSchema: EntitySchema = {
     },
   },
   create: {
-    preHandler: authMiddleware,
+    preHandler: authProfessor,
     schema: {
       security: [{ jwtAuth: [] }],
       body: createObraBodyValidation,
@@ -36,13 +40,14 @@ export const ObraSchema: EntitySchema = {
         201: genericSuccessResponse,
         400: schemaValidationError,
         401: genericError,
+        403: genericError,
         500: genericError,
       },
       summary: 'Cria nova obra',
     },
   },
   update: {
-    preHandler: authMiddleware,
+    preHandler: authProfessor,
     schema: {
       security: [{ jwtAuth: [] }],
       params: idValidation,
@@ -51,6 +56,7 @@ export const ObraSchema: EntitySchema = {
         200: genericSuccessResponse,
         400: schemaValidationError,
         401: genericError,
+        403: genericError,
         404: genericError,
         500: genericError,
       },
