+++ src/shared/schemas/UsuarioSchema.ts
@@ -1,4 +1,4 @@
-import { authUsuario } from '../plugins/auth';
+import { authMiddleware } from '../middlewares/Authentication';
 import { genericError, schemaValidationError } from '../Schemas';
 import type { EntitySchema } from '../Types';
 import { idValidation, genericSuccessResponse } from '../Validations';
@@ -11,7 +11,7 @@ import {
 
 export const UsuarioSchema: EntitySchema = {
   get: {
-    preHandler: authUsuario,
+    preHandler: authMiddleware,
     schema: {
       security: [{ jwtAuth: [] }],
       params: idValidation,
@@ -41,7 +41,7 @@ export const UsuarioSchema: EntitySchema = {
   },
 
   professorCreate: {
-    preHandler: authUsuario,
+    preHandler: authMiddleware,
     schema: {
       security: [{ jwtAuth: [] }],
       body: professorCreateBodyValidation,
@@ -56,7 +56,7 @@ export const UsuarioSchema: EntitySchema = {
   },
 
   update: {
-    preHandler: authUsuario,
+    preHandler: authMiddleware,
     schema: {
       security: [{ jwtAuth: [] }],
       params: idValidation,
@@ -73,7 +73,7 @@ export const UsuarioSchema: EntitySchema = {
   },
 
   delete: {
-    preHandler: authUsuario,
+    preHandler: authMiddleware,
     schema: {
       security: [{ jwtAuth: [] }],
       params: idValidation,
@@ -89,7 +89,7 @@ export const UsuarioSchema: EntitySchema = {
   },
 
   fotoCreate: {
-    preHandler: authUsuario,
+    preHandler: authMiddleware,
     schema: {
       security: [{ jwtAuth: [] }],
       params: idValidation,
@@ -106,7 +106,7 @@ export const UsuarioSchema: EntitySchema = {
   },
 
   fotoUpdate: {
-    preHandler: authUsuario,
+    preHandler: authMiddleware,
     schema: {
       security: [{ jwtAuth: [] }],
       params: idValidation,
@@ -122,7 +122,7 @@ export const UsuarioSchema: EntitySchema = {
   },
 
   fotoDelete: {
-    preHandler: authUsuario,
+    preHandler: authMiddleware,
     schema: {
       security: [{ jwtAuth: [] }],
       params: idValidation,
